Proper encoding of the payload in data:uri for ics link
When the event's payload contains special url characters, like this one: `#`,
if the data:uri is not properly encoded, the navigator will not correctly
navigate to the href: the data:uri will be parsed until the `#`.

The better solution in that case is to base64 encode the data:uri payload.